services/web/requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.4
Flask-CAS==1.0.0
Flask-login==0.5.0
passlib[bcrypt]==1.7
Flask-Mail
gunicorn==20.0.4
pandas==1.0.3
Flask-JSGlue
Werkzeug
geoalchemy2
flask_restful
SQLAlchemy-serializer
pyjwt
flask-crontab
requests
apscheduler

